+++ examples/adapter/paragraph.rs
@@ -1,5 +1,6 @@
-use crate::adapter::Outcome;
-use rat_event::{FocusKeys, HandleEvent, MouseOnly};
+use crate::adapter::_private::NonExhaustive;
+use rat_event::{FocusKeys, HandleEvent, MouseOnly, Outcome};
+use rat_scrolled::{ScrollingState, ScrollingWidget};
 use ratatui::buffer::Buffer;
 use ratatui::layout::{Alignment, Rect};
 use ratatui::prelude::{BlockExt, StatefulWidget};
@@ -7,8 +8,6 @@ use ratatui::style::Style;
 use ratatui::text::Text;
 use ratatui::widgets::{Block, Paragraph, Widget, Wrap};
 use std::cmp::min;
-use rat_scrolled::{ScrollingState, ScrollingWidget};
-use crate::adapter::_private::NonExhaustive;
 
 ///
 /// Adapter for ratatui::widget::Paragraph