+++ examples/slist.rs
@@ -12,7 +12,7 @@ use crossterm::terminal::{
 };
 use crossterm::ExecutableCommand;
 use log::debug;
-use rat_event::{HandleEvent, MouseOnly};
+use rat_event::{HandleEvent, MouseOnly, Outcome};
 use rat_scrolled::{Scrolled, ScrolledState};
 use ratatui::backend::CrosstermBackend;
 use ratatui::buffer::Buffer;
@@ -26,16 +26,6 @@ use std::iter::repeat_with;
 use std::time::Duration;
 
 pub mod adapter;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Outcome {
-    /// The given event was not handled at all.
-    NotUsed,
-    /// The event was handled, no repaint necessary.
-    Unchanged,
-    /// The event was handled, repaint necessary.
-    Changed,
-}
 
 fn main() -> Result<(), anyhow::Error> {
     setup_logging()?;
@@ -108,21 +98,6 @@ fn setup_logging() -> Result<(), anyhow::Error> {
         .chain(fern::log_file("log.log")?)
         .apply()?;
     Ok(())
-}
-
-impl From<rat_scrolled::event::Outcome<adapter::Outcome>> for Outcome {
-    fn from(value: rat_scrolled::event::Outcome<adapter::Outcome>) -> Self {
-        match value {
-            rat_scrolled::event::Outcome::Inner(i) => match i {
-                adapter::Outcome::NotUsed => Outcome::NotUsed,
-                adapter::Outcome::Unchanged => Outcome::Unchanged,
-                adapter::Outcome::Changed => Outcome::Changed,
-            },
-            rat_scrolled::event::Outcome::NotUsed => Outcome::NotUsed,
-            rat_scrolled::event::Outcome::Unchanged => Outcome::Unchanged,
-            rat_scrolled::event::Outcome::Changed => Outcome::Changed,
-        }
-    }
 }
 
 struct Data {