+++ examples/sparagraph.rs
@@ -11,6 +11,7 @@ use crossterm::terminal::{
     disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen,
 };
 use crossterm::ExecutableCommand;
+use rat_event::Outcome;
 use rat_scrolled::event::{HandleEvent, MouseOnly};
 use rat_scrolled::{Scrolled, ScrolledState};
 use ratatui::backend::CrosstermBackend;
@@ -23,31 +24,6 @@ use std::io::{stdout, Stdout};
 use std::time::Duration;
 
 pub mod adapter;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Outcome {
-    /// The given event was not handled at all.
-    NotUsed,
-    /// The event was handled, no repaint necessary.
-    Unchanged,
-    /// The event was handled, repaint necessary.
-    Changed,
-}
-
-impl From<rat_scrolled::event::Outcome<adapter::Outcome>> for Outcome {
-    fn from(value: rat_scrolled::event::Outcome<adapter::Outcome>) -> Self {
-        match value {
-            rat_scrolled::event::Outcome::Inner(i) => match i {
-                adapter::Outcome::NotUsed => Outcome::NotUsed,
-                adapter::Outcome::Unchanged => Outcome::Unchanged,
-                adapter::Outcome::Changed => Outcome::Changed,
-            },
-            rat_scrolled::event::Outcome::NotUsed => Outcome::NotUsed,
-            rat_scrolled::event::Outcome::Unchanged => Outcome::Unchanged,
-            rat_scrolled::event::Outcome::Changed => Outcome::Changed,
-        }
-    }
-}
 
 fn main() -> Result<(), anyhow::Error> {
     setup_logging()?;