+++ examples/stable.rs
@@ -11,7 +11,8 @@ use crossterm::terminal::{
     disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen,
 };
 use crossterm::ExecutableCommand;
-use rat_event::{HandleEvent, MouseOnly};
+use rat_event::{HandleEvent, MouseOnly, Outcome};
+use rat_scrolled::event::ScrollOutcome;
 use rat_scrolled::{Scrolled, ScrolledState};
 use ratatui::backend::CrosstermBackend;
 use ratatui::buffer::Buffer;
@@ -96,31 +97,6 @@ fn setup_logging() -> Result<(), anyhow::Error> {
         .chain(fern::log_file("log.log")?)
         .apply()?;
     Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Outcome {
-    /// The given event was not handled at all.
-    NotUsed,
-    /// The event was handled, no repaint necessary.
-    Unchanged,
-    /// The event was handled, repaint necessary.
-    Changed,
-}
-
-impl From<rat_scrolled::event::Outcome<adapter::Outcome>> for Outcome {
-    fn from(value: rat_scrolled::event::Outcome<adapter::Outcome>) -> Self {
-        match value {
-            rat_scrolled::event::Outcome::Inner(i) => match i {
-                adapter::Outcome::NotUsed => Outcome::NotUsed,
-                adapter::Outcome::Unchanged => Outcome::Unchanged,
-                adapter::Outcome::Changed => Outcome::Changed,
-            },
-            rat_scrolled::event::Outcome::NotUsed => Outcome::NotUsed,
-            rat_scrolled::event::Outcome::Unchanged => Outcome::Unchanged,
-            rat_scrolled::event::Outcome::Changed => Outcome::Changed,
-        }
-    }
 }
 
 struct Data {